Updated models_psql.py.

+++ domrf_scrapy/domrf_scrapy/models_psql.py
@@ -67,8 +67,8 @@ class Region(Base_item):
 class DeveloperData(Base_item):
     __tablename__ = "developer_Data"
     id = Column(BIGINT, primary_key = True)
-    developer_group_id = Column(INTEGER, ForeignKey("developers.developer_group_id"), nullable=False)
-    region_id = Column(INTEGER, ForeignKey("regions.region_id"), nullable=False)
+    developer_group_id = Column(BIGINT, ForeignKey("developers.developer_group_id"), nullable=False)
+    region_id = Column(BIGINT, ForeignKey("regions.region_id"), nullable=False)
     startDate = Column(VARCHAR(20))
     endDate = Column(VARCHAR(20))
